+++ test/unit/Price.ts
@@ -17,20 +17,9 @@ describe("#Price", () => {
 
   describe("#current", () => {
     describe("#single currency", () => {
-      let sandbox: any
-      beforeEach(() => (sandbox = sinon.sandbox.create()))
-      afterEach(() => sandbox.restore())
-
-      it("should get current price for single currency", done => {
-        const data: any = []
-        const resolved = new Promise(r => r({ data: data }))
-        sandbox.stub(axios, "get").returns(resolved)
-
-        bitbox.Price.current("usd")
-          .then((result: any) => {
-            assert.deepEqual(data.price, result)
-          })
-          .then(done, done)
+      it("should get current price for single currency", async () => {
+        const result = await bitbox.Price.current("usd")
+        assert.notEqual(0, result)
       })
     })
   })
