+++ test/unit/RawTransactions.js
@@ -42,9 +42,8 @@ Object.defineProperty(exports, "__esModule", { value: true });
    See the sendRawTransaction test for an example.
   -Create a mocking library of data to compare unit and integration tests.
 */
-var assert = require("assert");
-var axios_1 = require("axios");
-var sinon = require("sinon");
+var chai = require("chai");
+var assert = chai.assert;
 // TODO: port from require to import syntax
 var BITBOX = require("../../lib/BITBOX").BITBOX;
 var bitbox = new BITBOX();
@@ -55,7 +54,7 @@ var nock = require("nock"); // HTTP mocking
 var util = require("util");
 util.inspect.defaultOptions = { depth: 1 };
 describe("#RawTransactions", function () {
-    describe("#PriceConstructor", function () {
+    describe("#RawTransactionsConstructor", function () {
         it("should create instance of RawTransactions", function () {
             var rawtransactions = new RawTransactions();
             assert.equal(rawtransactions instanceof RawTransactions, true);
@@ -66,110 +65,396 @@ describe("#RawTransactions", function () {
         });
     });
     describe("#decodeRawTransaction", function () {
-        var sandbox;
-        beforeEach(function () { return (sandbox = sinon.sandbox.create()); });
-        afterEach(function () { return sandbox.restore(); });
-        it("should decode raw transaction", function (done) {
-            var data = {
-                txid: "4ebd325a4b394cff8c57e8317ccf5a8d0e2bdf1b8526f8aad6c8e43d8240621a",
-                hash: "4ebd325a4b394cff8c57e8317ccf5a8d0e2bdf1b8526f8aad6c8e43d8240621a",
-                size: 10,
-                version: 2,
-                locktime: 0,
-                vin: [],
-                vout: []
-            };
-            var resolved = new Promise(function (r) { return r({ data: data }); });
-            sandbox.stub(axios_1.default, "get").returns(resolved);
-            bitbox.RawTransactions.decodeRawTransaction(["02000000000000000000"])
-                .then(function (result) {
-                assert.deepEqual(data, result[0]);
-            })
-                .then(done, done);
-        });
-    });
-    describe("#decodeScript", function () {
-        var sandbox;
-        beforeEach(function () { return (sandbox = sinon.sandbox.create()); });
-        afterEach(function () { return sandbox.restore(); });
-        it("should decode script", function () { return __awaiter(_this, void 0, void 0, function () {
-            var data, resolved, result;
+        it("should decode tx for a single hex", function () { return __awaiter(_this, void 0, void 0, function () {
+            var hex, result;
             return __generator(this, function (_a) {
                 switch (_a.label) {
                     case 0:
-                        data = {
-                            asm: "OP_RETURN 5361746f736869204e616b616d6f746f",
-                            type: "nulldata",
-                            p2sh: "bitcoincash:prswx5965nfumux9qng5kj8hw603vcne7q08t8c6jp"
-                        };
-                        resolved = new Promise(function (r) { return r({ data: data }); });
-                        sandbox.stub(axios_1.default, "get").returns(resolved);
-                        return [4 /*yield*/, bitbox.RawTransactions.decodeScript("6a105361746f736869204e616b616d6f746f")
-                            //console.log(`result: ${util.inspect(result)}`)
+                        hex = "0200000001b9b598d7d6d72fc486b2b3a3c03c79b5bade6ec9a77ced850515ab5e64edcc21010000006b483045022100a7b1b08956abb8d6f322aa709d8583c8ea492ba0585f1a6f4f9983520af74a5a0220411aee4a9a54effab617b0508c504c31681b15f9b187179b4874257badd4139041210360cfc66fdacb650bc4c83b4e351805181ee696b7d5ab4667c57b2786f51c413dffffffff0210270000000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac786e9800000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac00000000";
+                        return [4 /*yield*/, bitbox.RawTransactions.decodeRawTransaction(hex)
+                            //console.log(`result ${JSON.stringify(result, null, 2)}`)
                         ];
                     case 1:
                         result = _a.sent();
-                        //console.log(`result: ${util.inspect(result)}`)
-                        assert.deepEqual(data, result);
+                        //console.log(`result ${JSON.stringify(result, null, 2)}`)
+                        assert.hasAnyKeys(result, [
+                            "txid",
+                            "hash",
+                            "size",
+                            "version",
+                            "locktime",
+                            "vin",
+                            "vout"
+                        ]);
+                        assert.isArray(result.vin);
+                        assert.isArray(result.vout);
                         return [2 /*return*/];
+                }
+            });
+        }); });
+        it("should decode an array of tx hexes", function () { return __awaiter(_this, void 0, void 0, function () {
+            var hexes, result;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        hexes = [
+                            "0200000001b9b598d7d6d72fc486b2b3a3c03c79b5bade6ec9a77ced850515ab5e64edcc21010000006b483045022100a7b1b08956abb8d6f322aa709d8583c8ea492ba0585f1a6f4f9983520af74a5a0220411aee4a9a54effab617b0508c504c31681b15f9b187179b4874257badd4139041210360cfc66fdacb650bc4c83b4e351805181ee696b7d5ab4667c57b2786f51c413dffffffff0210270000000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac786e9800000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac00000000",
+                            "0200000001b9b598d7d6d72fc486b2b3a3c03c79b5bade6ec9a77ced850515ab5e64edcc21010000006b483045022100a7b1b08956abb8d6f322aa709d8583c8ea492ba0585f1a6f4f9983520af74a5a0220411aee4a9a54effab617b0508c504c31681b15f9b187179b4874257badd4139041210360cfc66fdacb650bc4c83b4e351805181ee696b7d5ab4667c57b2786f51c413dffffffff0210270000000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac786e9800000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac00000000"
+                        ];
+                        return [4 /*yield*/, bitbox.RawTransactions.decodeRawTransaction(hexes)
+                            //console.log(`result ${JSON.stringify(result, null, 2)}`)
+                        ];
+                    case 1:
+                        result = _a.sent();
+                        //console.log(`result ${JSON.stringify(result, null, 2)}`)
+                        assert.isArray(result);
+                        assert.hasAnyKeys(result[0], [
+                            "txid",
+                            "hash",
+                            "size",
+                            "version",
+                            "locktime",
+                            "vin",
+                            "vout"
+                        ]);
+                        assert.isArray(result[0].vin);
+                        assert.isArray(result[0].vout);
+                        return [2 /*return*/];
+                }
+            });
+        }); });
+        it("should throw an error for improper single input", function () { return __awaiter(_this, void 0, void 0, function () {
+            var addr, err_1;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        _a.trys.push([0, 2, , 3]);
+                        addr = 12345;
+                        return [4 /*yield*/, bitbox.RawTransactions.decodeRawTransaction(addr)];
+                    case 1:
+                        _a.sent();
+                        assert.equal(true, false, "Unexpected result!");
+                        return [3 /*break*/, 3];
+                    case 2:
+                        err_1 = _a.sent();
+                        //console.log(`err: `, err)
+                        assert.include(err_1.message, "Input must be a string or array of strings.");
+                        return [3 /*break*/, 3];
+                    case 3: return [2 /*return*/];
+                }
+            });
+        }); });
+        it("should throw error on array size rate limit", function () { return __awaiter(_this, void 0, void 0, function () {
+            var data, i, result, err_2;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        _a.trys.push([0, 2, , 3]);
+                        data = [];
+                        for (i = 0; i < 25; i++) {
+                            data.push("0200000001b9b598d7d6d72fc486b2b3a3c03c79b5bade6ec9a77ced850515ab5e64edcc21010000006b483045022100a7b1b08956abb8d6f322aa709d8583c8ea492ba0585f1a6f4f9983520af74a5a0220411aee4a9a54effab617b0508c504c31681b15f9b187179b4874257badd4139041210360cfc66fdacb650bc4c83b4e351805181ee696b7d5ab4667c57b2786f51c413dffffffff0210270000000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac786e9800000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac00000000");
+                        }
+                        return [4 /*yield*/, bitbox.RawTransactions.decodeRawTransaction(data)];
+                    case 1:
+                        result = _a.sent();
+                        console.log("result: " + util.inspect(result));
+                        assert.equal(true, false, "Unexpected result!");
+                        return [3 /*break*/, 3];
+                    case 2:
+                        err_2 = _a.sent();
+                        assert.hasAnyKeys(err_2, ["error"]);
+                        assert.include(err_2.error, "Array too large");
+                        return [3 /*break*/, 3];
+                    case 3: return [2 /*return*/];
                 }
             });
         }); });
     });
     describe("#getRawTransaction", function () {
-        var sandbox;
-        beforeEach(function () { return (sandbox = sinon.sandbox.create()); });
-        afterEach(function () { return sandbox.restore(); });
-        it("should get raw transaction", function (done) {
-            var data = "020000000160d663961c63c7f0a07f22ec07b8f55b3935bfdbed8b1d8454916e8932fbf109010000006b4830450221008479fab4cfdcb111833d250a43f98ac26d43272b7a29cb1b9a0491eae5c44b3502203448b17253632395c29a7d62058bbfe93efb20fc8636ba6837002d464195aec04121029123258f7cdcd45b864066bcaa9b71f24d5ed1fa1dd36eaf107d8432b5014658ffffffff016d180000000000001976a91479d3297d1823149f4ec61df31d19f2fad5390c0288ac00000000";
-            var resolved = new Promise(function (r) { return r({ data: data }); });
-            sandbox.stub(axios_1.default, "get").returns(resolved);
-            bitbox.RawTransactions.getRawTransaction([
-                "808d617eccaad4f1397fe07a06ec5ed15a0821cf22a3e0931c0c92aef9e572b6"
-            ])
-                .then(function (result) {
-                assert.deepEqual(data, result[0]);
-            })
-                .then(done, done);
-        });
-    });
-    describe("#sendRawTransaction", function () {
-        var sandbox;
-        beforeEach(function () { return (sandbox = sinon.sandbox.create()); });
-        afterEach(function () { return sandbox.restore(); });
-        it("should send single raw transaction", function () { return __awaiter(_this, void 0, void 0, function () {
-            var data, resolved, result;
+        it("should decode a single txid, with concise output", function () { return __awaiter(_this, void 0, void 0, function () {
+            var txid, verbose, result;
             return __generator(this, function (_a) {
                 switch (_a.label) {
                     case 0:
-                        data = "Error: transaction already in block chain";
-                        resolved = new Promise(function (r) { return r({ data: data }); });
-                        sandbox.stub(axios_1.default, "get").returns(resolved);
-                        return [4 /*yield*/, bitbox.RawTransactions.sendRawTransaction("020000000160d663961c63c7f0a07f22ec07b8f55b3935bfdbed8b1d8454916e8932fbf109010000006b4830450221008479fab4cfdcb111833d250a43f98ac26d43272b7a29cb1b9a0491eae5c44b3502203448b17253632395c29a7d62058bbfe93efb20fc8636ba6837002d464195aec04121029123258f7cdcd45b864066bcaa9b71f24d5ed1fa1dd36eaf107d8432b5014658ffffffff016d180000000000001976a91479d3297d1823149f4ec61df31d19f2fad5390c0288ac00000000")];
+                        txid = "23213453b4642a73b4fc30d3112d72549ca153a8707255b14373b59e43558de1";
+                        verbose = false;
+                        return [4 /*yield*/, bitbox.RawTransactions.getRawTransaction(txid, verbose)
+                            //console.log(`result: ${JSON.stringify(result, null, 2)}`)
+                        ];
                     case 1:
                         result = _a.sent();
-                        assert.equal(data, result);
+                        //console.log(`result: ${JSON.stringify(result, null, 2)}`)
+                        assert.isString(result);
                         return [2 /*return*/];
                 }
             });
         }); });
-        it("should send an array of raw transactions", function () { return __awaiter(_this, void 0, void 0, function () {
-            var data, result;
+        it("should decode a single txid, with verbose output", function () { return __awaiter(_this, void 0, void 0, function () {
+            var txid, verbose, result;
             return __generator(this, function (_a) {
                 switch (_a.label) {
                     case 0:
-                        data = "Error: transaction already in block chain";
-                        // Mock the http call to rest.bitcoin.com
-                        nock("" + bitbox.RawTransactions.restURL)
-                            .post(function (uri) { return uri.includes("/"); })
-                            .reply(200, { data: data });
-                        return [4 /*yield*/, bitbox.RawTransactions.sendRawTransaction([
-                                "020000000160d663961c63c7f0a07f22ec07b8f55b3935bfdbed8b1d8454916e8932fbf109010000006b4830450221008479fab4cfdcb111833d250a43f98ac26d43272b7a29cb1b9a0491eae5c44b3502203448b17253632395c29a7d62058bbfe93efb20fc8636ba6837002d464195aec04121029123258f7cdcd45b864066bcaa9b71f24d5ed1fa1dd36eaf107d8432b5014658ffffffff016d180000000000001976a91479d3297d1823149f4ec61df31d19f2fad5390c0288ac00000000"
-                            ])];
+                        txid = "23213453b4642a73b4fc30d3112d72549ca153a8707255b14373b59e43558de1";
+                        verbose = true;
+                        return [4 /*yield*/, bitbox.RawTransactions.getRawTransaction(txid, verbose)
+                            //console.log(`result: ${JSON.stringify(result, null, 2)}`)
+                        ];
                     case 1:
                         result = _a.sent();
-                        assert.equal(data, result.data);
+                        //console.log(`result: ${JSON.stringify(result, null, 2)}`)
+                        assert.hasAnyKeys(result, [
+                            "hex",
+                            "txid",
+                            "hash",
+                            "size",
+                            "version",
+                            "locktime",
+                            "vin",
+                            "vout",
+                            "blockhash",
+                            "confirmations",
+                            "time",
+                            "blocktime"
+                        ]);
+                        assert.isArray(result.vin);
+                        assert.isArray(result.vout);
                         return [2 /*return*/];
+                }
+            });
+        }); });
+        it("should decode an array of txids, with a concise output", function () { return __awaiter(_this, void 0, void 0, function () {
+            var txid, verbose, result;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        txid = [
+                            "23213453b4642a73b4fc30d3112d72549ca153a8707255b14373b59e43558de1",
+                            "b25d24fbb42d84812ed2cb55797f10fdec41afc7906ab563d1ec8c8676a2037f"
+                        ];
+                        verbose = false;
+                        return [4 /*yield*/, bitbox.RawTransactions.getRawTransaction(txid, verbose)
+                            //console.log(`result: ${JSON.stringify(result, null, 2)}`)
+                        ];
+                    case 1:
+                        result = _a.sent();
+                        //console.log(`result: ${JSON.stringify(result, null, 2)}`)
+                        assert.isArray(result);
+                        assert.isString(result[0]);
+                        return [2 /*return*/];
+                }
+            });
+        }); });
+        it("should decode an array of txids, with a verbose output", function () { return __awaiter(_this, void 0, void 0, function () {
+            var txid, verbose, result;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        txid = [
+                            "23213453b4642a73b4fc30d3112d72549ca153a8707255b14373b59e43558de1",
+                            "b25d24fbb42d84812ed2cb55797f10fdec41afc7906ab563d1ec8c8676a2037f"
+                        ];
+                        verbose = true;
+                        return [4 /*yield*/, bitbox.RawTransactions.getRawTransaction(txid, verbose)
+                            //console.log(`result: ${JSON.stringify(result, null, 2)}`)
+                        ];
+                    case 1:
+                        result = _a.sent();
+                        //console.log(`result: ${JSON.stringify(result, null, 2)}`)
+                        assert.isArray(result);
+                        assert.hasAnyKeys(result[0], [
+                            "hex",
+                            "txid",
+                            "hash",
+                            "size",
+                            "version",
+                            "locktime",
+                            "vin",
+                            "vout",
+                            "blockhash",
+                            "confirmations",
+                            "time",
+                            "blocktime"
+                        ]);
+                        assert.isArray(result[0].vin);
+                        assert.isArray(result[0].vout);
+                        return [2 /*return*/];
+                }
+            });
+        }); });
+        it("should throw error on array size rate limit", function () { return __awaiter(_this, void 0, void 0, function () {
+            var dataMock, data, i, result, err_3;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        _a.trys.push([0, 2, , 3]);
+                        dataMock = "23213453b4642a73b4fc30d3112d72549ca153a8707255b14373b59e43558de1";
+                        data = [];
+                        for (i = 0; i < 25; i++)
+                            data.push(dataMock);
+                        return [4 /*yield*/, bitbox.RawTransactions.getRawTransaction(data)];
+                    case 1:
+                        result = _a.sent();
+                        console.log("result: " + util.inspect(result));
+                        assert.equal(true, false, "Unexpected result!");
+                        return [3 /*break*/, 3];
+                    case 2:
+                        err_3 = _a.sent();
+                        assert.hasAnyKeys(err_3, ["error"]);
+                        assert.include(err_3.error, "Array too large");
+                        return [3 /*break*/, 3];
+                    case 3: return [2 /*return*/];
+                }
+            });
+        }); });
+    });
+    describe("#decodeScript", function () {
+        it("should decode script for a single hex", function () { return __awaiter(_this, void 0, void 0, function () {
+            var hex, result;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        hex = "4830450221009a51e00ec3524a7389592bc27bea4af5104a59510f5f0cfafa64bbd5c164ca2e02206c2a8bbb47eabdeed52f17d7df668d521600286406930426e3a9415fe10ed592012102e6e1423f7abde8b70bca3e78a7d030e5efabd3eb35c19302542b5fe7879c1a16";
+                        return [4 /*yield*/, bitbox.RawTransactions.decodeScript(hex)
+                            //console.log(`result ${JSON.stringify(result, null, 2)}`)
+                        ];
+                    case 1:
+                        result = _a.sent();
+                        //console.log(`result ${JSON.stringify(result, null, 2)}`)
+                        assert.hasAllKeys(result, ["asm", "type", "p2sh"]);
+                        return [2 /*return*/];
+                }
+            });
+        }); });
+        // CT 2/20/19 - Waiting for this PR to be merged complete the test:
+        // https://github.com/Bitcoin-com/rest.bitcoin.com/pull/312
+        /*
+        it("should decode an array of tx hexes", async () => {
+          const hexes = [
+            "4830450221009a51e00ec3524a7389592bc27bea4af5104a59510f5f0cfafa64bbd5c164ca2e02206c2a8bbb47eabdeed52f17d7df668d521600286406930426e3a9415fe10ed592012102e6e1423f7abde8b70bca3e78a7d030e5efabd3eb35c19302542b5fe7879c1a16",
+            "4830450221009a51e00ec3524a7389592bc27bea4af5104a59510f5f0cfafa64bbd5c164ca2e02206c2a8bbb47eabdeed52f17d7df668d521600286406930426e3a9415fe10ed592012102e6e1423f7abde8b70bca3e78a7d030e5efabd3eb35c19302542b5fe7879c1a16"
+          ]
+    
+          const result = await bitbox.RawTransactions.decodeScript(hexes)
+          console.log(`result ${JSON.stringify(result, null, 2)}`)
+        })
+    */
+        /*
+        it(`should throw an error for improper single input`, async () => {
+          try {
+            const addr = 12345
+    
+            await bitbox.RawTransactions.decodeRawTransaction(addr)
+            assert.equal(true, false, "Unexpected result!")
+          } catch (err) {
+            //console.log(`err: `, err)
+            assert.include(
+              err.message,
+              `Input must be a string or array of strings.`
+            )
+          }
+        })
+    */
+    });
+    /*
+      Testing sentRawTransaction isn't really possible with an integration test,
+      as the endpoint really needs an e2e test to be properly tested. The tests
+      below expect error messages returned from the server, but at least test
+      that the server is responding on those endpoints, and responds consistently.
+    */
+    describe("sendRawTransaction", function () {
+        it("should send a single transaction hex", function () { return __awaiter(_this, void 0, void 0, function () {
+            var hex, err_4;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        _a.trys.push([0, 2, , 3]);
+                        hex = "01000000013ba3edfd7a7b12b27ac72c3e67768f617fc81bc3888a51323a9fb8aa4b1e5e4a000000006a4730440220540986d1c58d6e76f8f05501c520c38ce55393d0ed7ed3c3a82c69af04221232022058ea43ed6c05fec0eccce749a63332ed4525460105346f11108b9c26df93cd72012103083dfc5a0254613941ddc91af39ff90cd711cdcde03a87b144b883b524660c39ffffffff01807c814a000000001976a914d7e7c4e0b70eaa67ceff9d2823d1bbb9f6df9a5188ac00000000";
+                        return [4 /*yield*/, bitbox.RawTransactions.sendRawTransaction(hex)
+                            //console.log(`result ${JSON.stringify(result, null, 2)}`)
+                        ];
+                    case 1:
+                        _a.sent();
+                        //console.log(`result ${JSON.stringify(result, null, 2)}`)
+                        assert.equal(true, false, "Unexpected result!");
+                        return [3 /*break*/, 3];
+                    case 2:
+                        err_4 = _a.sent();
+                        //console.log(`err: ${util.inspect(err)}`)
+                        assert.hasAllKeys(err_4, ["error"]);
+                        assert.include(err_4.error, "Missing inputs");
+                        return [3 /*break*/, 3];
+                    case 3: return [2 /*return*/];
+                }
+            });
+        }); });
+        it("should send an array of tx hexes", function () { return __awaiter(_this, void 0, void 0, function () {
+            var hexes, result, err_5;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        _a.trys.push([0, 2, , 3]);
+                        hexes = [
+                            "01000000013ba3edfd7a7b12b27ac72c3e67768f617fc81bc3888a51323a9fb8aa4b1e5e4a000000006a4730440220540986d1c58d6e76f8f05501c520c38ce55393d0ed7ed3c3a82c69af04221232022058ea43ed6c05fec0eccce749a63332ed4525460105346f11108b9c26df93cd72012103083dfc5a0254613941ddc91af39ff90cd711cdcde03a87b144b883b524660c39ffffffff01807c814a000000001976a914d7e7c4e0b70eaa67ceff9d2823d1bbb9f6df9a5188ac00000000",
+                            "01000000013ba3edfd7a7b12b27ac72c3e67768f617fc81bc3888a51323a9fb8aa4b1e5e4a000000006a4730440220540986d1c58d6e76f8f05501c520c38ce55393d0ed7ed3c3a82c69af04221232022058ea43ed6c05fec0eccce749a63332ed4525460105346f11108b9c26df93cd72012103083dfc5a0254613941ddc91af39ff90cd711cdcde03a87b144b883b524660c39ffffffff01807c814a000000001976a914d7e7c4e0b70eaa67ceff9d2823d1bbb9f6df9a5188ac00000000"
+                        ];
+                        return [4 /*yield*/, bitbox.RawTransactions.sendRawTransaction(hexes)];
+                    case 1:
+                        result = _a.sent();
+                        console.log("result " + JSON.stringify(result, null, 2));
+                        return [3 /*break*/, 3];
+                    case 2:
+                        err_5 = _a.sent();
+                        // console.log(`err: ${util.inspect(err)}`)
+                        assert.hasAllKeys(err_5, ["error"]);
+                        assert.include(err_5.error, "Missing inputs");
+                        return [3 /*break*/, 3];
+                    case 3: return [2 /*return*/];
+                }
+            });
+        }); });
+        it("should throw an error for improper single input", function () { return __awaiter(_this, void 0, void 0, function () {
+            var addr, err_6;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        _a.trys.push([0, 2, , 3]);
+                        addr = 12345;
+                        return [4 /*yield*/, bitbox.RawTransactions.sendRawTransaction(addr)];
+                    case 1:
+                        _a.sent();
+                        assert.equal(true, false, "Unexpected result!");
+                        return [3 /*break*/, 3];
+                    case 2:
+                        err_6 = _a.sent();
+                        //console.log(`err: `, err)
+                        assert.include(err_6.message, "Input hex must be a string or array of strings");
+                        return [3 /*break*/, 3];
+                    case 3: return [2 /*return*/];
+                }
+            });
+        }); });
+        it("should throw error on array size rate limit", function () { return __awaiter(_this, void 0, void 0, function () {
+            var dataMock, data, i, result, err_7;
+            return __generator(this, function (_a) {
+                switch (_a.label) {
+                    case 0:
+                        _a.trys.push([0, 2, , 3]);
+                        dataMock = "01000000013ba3edfd7a7b12b27ac72c3e67768f617fc81bc3888a51323a9fb8aa4b1e5e4a000000006a4730440220540986d1c58d6e76f8f05501c520c38ce55393d0ed7ed3c3a82c69af04221232022058ea43ed6c05fec0eccce749a63332ed4525460105346f11108b9c26df93cd72012103083dfc5a0254613941ddc91af39ff90cd711cdcde03a87b144b883b524660c39ffffffff01807c814a000000001976a914d7e7c4e0b70eaa67ceff9d2823d1bbb9f6df9a5188ac00000000";
+                        data = [];
+                        for (i = 0; i < 25; i++)
+                            data.push(dataMock);
+                        return [4 /*yield*/, bitbox.RawTransactions.sendRawTransaction(data)];
+                    case 1:
+                        result = _a.sent();
+                        console.log("result: " + util.inspect(result));
+                        assert.equal(true, false, "Unexpected result!");
+                        return [3 /*break*/, 3];
+                    case 2:
+                        err_7 = _a.sent();
+                        assert.hasAnyKeys(err_7, ["error"]);
+                        assert.include(err_7.error, "Array too large");
+                        return [3 /*break*/, 3];
+                    case 3: return [2 /*return*/];
                 }
             });
         }); });
