+++ test/unit/RawTransactions.ts
@@ -4,9 +4,8 @@
    See the sendRawTransaction test for an example.
   -Create a mocking library of data to compare unit and integration tests.
 */
-import * as assert from "assert";
-import axios from "axios";
-import * as sinon from "sinon";
+import * as chai from "chai"
+const assert = chai.assert
 
 // TODO: port from require to import syntax
 const BITBOX = require("../../lib/BITBOX").BITBOX
@@ -20,7 +19,7 @@ const util = require("util")
 util.inspect.defaultOptions = { depth: 1 }
 
 describe("#RawTransactions", () => {
-  describe("#PriceConstructor", () => {
+  describe("#RawTransactionsConstructor", () => {
     it("should create instance of RawTransactions", () => {
       const rawtransactions = new RawTransactions()
       assert.equal(rawtransactions instanceof RawTransactions, true)
@@ -33,111 +32,308 @@ describe("#RawTransactions", () => {
   })
 
   describe("#decodeRawTransaction", () => {
-    let sandbox: any
-    beforeEach(() => (sandbox = sinon.sandbox.create()))
-    afterEach(() => sandbox.restore())
+    it("should decode tx for a single hex", async () => {
+      const hex =
+        "0200000001b9b598d7d6d72fc486b2b3a3c03c79b5bade6ec9a77ced850515ab5e64edcc21010000006b483045022100a7b1b08956abb8d6f322aa709d8583c8ea492ba0585f1a6f4f9983520af74a5a0220411aee4a9a54effab617b0508c504c31681b15f9b187179b4874257badd4139041210360cfc66fdacb650bc4c83b4e351805181ee696b7d5ab4667c57b2786f51c413dffffffff0210270000000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac786e9800000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac00000000"
 
-    it("should decode raw transaction", done => {
-      const data = {
-        txid:
-          "4ebd325a4b394cff8c57e8317ccf5a8d0e2bdf1b8526f8aad6c8e43d8240621a",
-        hash:
-          "4ebd325a4b394cff8c57e8317ccf5a8d0e2bdf1b8526f8aad6c8e43d8240621a",
-        size: 10,
-        version: 2,
-        locktime: 0,
-        vin: [],
-        vout: []
-      }
+      const result = await bitbox.RawTransactions.decodeRawTransaction(hex)
+      //console.log(`result ${JSON.stringify(result, null, 2)}`)
 
-      const resolved = new Promise(r => r({ data: data }))
-      sandbox.stub(axios, "get").returns(resolved)
-
-      bitbox.RawTransactions.decodeRawTransaction(["02000000000000000000"])
-        .then((result: any) => {
-          assert.deepEqual(data, result[0])
-        })
-        .then(done, done)
+      assert.hasAnyKeys(result, [
+        "txid",
+        "hash",
+        "size",
+        "version",
+        "locktime",
+        "vin",
+        "vout"
+      ])
+      assert.isArray(result.vin)
+      assert.isArray(result.vout)
     })
-  })
 
-  describe("#decodeScript", () => {
-    let sandbox: any
-    beforeEach(() => (sandbox = sinon.sandbox.create()))
-    afterEach(() => sandbox.restore())
+    it("should decode an array of tx hexes", async () => {
+      const hexes = [
+        "0200000001b9b598d7d6d72fc486b2b3a3c03c79b5bade6ec9a77ced850515ab5e64edcc21010000006b483045022100a7b1b08956abb8d6f322aa709d8583c8ea492ba0585f1a6f4f9983520af74a5a0220411aee4a9a54effab617b0508c504c31681b15f9b187179b4874257badd4139041210360cfc66fdacb650bc4c83b4e351805181ee696b7d5ab4667c57b2786f51c413dffffffff0210270000000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac786e9800000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac00000000",
+        "0200000001b9b598d7d6d72fc486b2b3a3c03c79b5bade6ec9a77ced850515ab5e64edcc21010000006b483045022100a7b1b08956abb8d6f322aa709d8583c8ea492ba0585f1a6f4f9983520af74a5a0220411aee4a9a54effab617b0508c504c31681b15f9b187179b4874257badd4139041210360cfc66fdacb650bc4c83b4e351805181ee696b7d5ab4667c57b2786f51c413dffffffff0210270000000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac786e9800000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac00000000"
+      ]
 
-    it("should decode script", async () => {
-      const data = {
-        asm: "OP_RETURN 5361746f736869204e616b616d6f746f",
-        type: "nulldata",
-        p2sh: "bitcoincash:prswx5965nfumux9qng5kj8hw603vcne7q08t8c6jp"
+      const result = await bitbox.RawTransactions.decodeRawTransaction(hexes)
+      //console.log(`result ${JSON.stringify(result, null, 2)}`)
+
+      assert.isArray(result)
+      assert.hasAnyKeys(result[0], [
+        "txid",
+        "hash",
+        "size",
+        "version",
+        "locktime",
+        "vin",
+        "vout"
+      ])
+      assert.isArray(result[0].vin)
+      assert.isArray(result[0].vout)
+    })
+
+    it(`should throw an error for improper single input`, async () => {
+      try {
+        const addr = 12345
+
+        await bitbox.RawTransactions.decodeRawTransaction(addr)
+        assert.equal(true, false, "Unexpected result!")
+      } catch (err) {
+        //console.log(`err: `, err)
+        assert.include(
+          err.message,
+          `Input must be a string or array of strings.`
+        )
       }
+    })
 
-      const resolved = new Promise(r => r({ data: data }))
-      sandbox.stub(axios, "get").returns(resolved)
+    it(`should throw error on array size rate limit`, async () => {
+      try {
+        const data = []
+        for (let i = 0; i < 25; i++) {
+          data.push(
+            "0200000001b9b598d7d6d72fc486b2b3a3c03c79b5bade6ec9a77ced850515ab5e64edcc21010000006b483045022100a7b1b08956abb8d6f322aa709d8583c8ea492ba0585f1a6f4f9983520af74a5a0220411aee4a9a54effab617b0508c504c31681b15f9b187179b4874257badd4139041210360cfc66fdacb650bc4c83b4e351805181ee696b7d5ab4667c57b2786f51c413dffffffff0210270000000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac786e9800000000001976a914eb4b180def88e3f5625b2d8ae2c098ff7d85f66488ac00000000"
+          )
+        }
 
-      const result = await bitbox.RawTransactions.decodeScript(
-        "6a105361746f736869204e616b616d6f746f"
-      )
-      //console.log(`result: ${util.inspect(result)}`)
+        const result = await bitbox.RawTransactions.decodeRawTransaction(data)
 
-      assert.deepEqual(data, result)
+        console.log(`result: ${util.inspect(result)}`)
+        assert.equal(true, false, "Unexpected result!")
+      } catch (err) {
+        assert.hasAnyKeys(err, ["error"])
+        assert.include(err.error, "Array too large")
+      }
     })
   })
 
   describe("#getRawTransaction", () => {
-    let sandbox: any
-    beforeEach(() => (sandbox = sinon.sandbox.create()))
-    afterEach(() => sandbox.restore())
+    it("should decode a single txid, with concise output", async () => {
+      const txid =
+        "23213453b4642a73b4fc30d3112d72549ca153a8707255b14373b59e43558de1"
+      const verbose = false
 
-    it("should get raw transaction", done => {
-      const data =
-        "020000000160d663961c63c7f0a07f22ec07b8f55b3935bfdbed8b1d8454916e8932fbf109010000006b4830450221008479fab4cfdcb111833d250a43f98ac26d43272b7a29cb1b9a0491eae5c44b3502203448b17253632395c29a7d62058bbfe93efb20fc8636ba6837002d464195aec04121029123258f7cdcd45b864066bcaa9b71f24d5ed1fa1dd36eaf107d8432b5014658ffffffff016d180000000000001976a91479d3297d1823149f4ec61df31d19f2fad5390c0288ac00000000"
+      const result = await bitbox.RawTransactions.getRawTransaction(
+        txid,
+        verbose
+      )
+      //console.log(`result: ${JSON.stringify(result, null, 2)}`)
 
-      const resolved = new Promise(r => r({ data: data }))
-      sandbox.stub(axios, "get").returns(resolved)
+      assert.isString(result)
+    })
 
-      bitbox.RawTransactions.getRawTransaction([
-        "808d617eccaad4f1397fe07a06ec5ed15a0821cf22a3e0931c0c92aef9e572b6"
+    it("should decode a single txid, with verbose output", async () => {
+      const txid =
+        "23213453b4642a73b4fc30d3112d72549ca153a8707255b14373b59e43558de1"
+      const verbose = true
+
+      const result = await bitbox.RawTransactions.getRawTransaction(
+        txid,
+        verbose
+      )
+      //console.log(`result: ${JSON.stringify(result, null, 2)}`)
+
+      assert.hasAnyKeys(result, [
+        "hex",
+        "txid",
+        "hash",
+        "size",
+        "version",
+        "locktime",
+        "vin",
+        "vout",
+        "blockhash",
+        "confirmations",
+        "time",
+        "blocktime"
       ])
-        .then((result: any) => {
-          assert.deepEqual(data, result[0])
-        })
-        .then(done, done)
+      assert.isArray(result.vin)
+      assert.isArray(result.vout)
+    })
+
+    it("should decode an array of txids, with a concise output", async () => {
+      const txid = [
+        "23213453b4642a73b4fc30d3112d72549ca153a8707255b14373b59e43558de1",
+        "b25d24fbb42d84812ed2cb55797f10fdec41afc7906ab563d1ec8c8676a2037f"
+      ]
+      const verbose = false
+
+      const result = await bitbox.RawTransactions.getRawTransaction(
+        txid,
+        verbose
+      )
+      //console.log(`result: ${JSON.stringify(result, null, 2)}`)
+
+      assert.isArray(result)
+      assert.isString(result[0])
+    })
+
+    it("should decode an array of txids, with a verbose output", async () => {
+      const txid = [
+        "23213453b4642a73b4fc30d3112d72549ca153a8707255b14373b59e43558de1",
+        "b25d24fbb42d84812ed2cb55797f10fdec41afc7906ab563d1ec8c8676a2037f"
+      ]
+      const verbose = true
+
+      const result = await bitbox.RawTransactions.getRawTransaction(
+        txid,
+        verbose
+      )
+      //console.log(`result: ${JSON.stringify(result, null, 2)}`)
+
+      assert.isArray(result)
+      assert.hasAnyKeys(result[0], [
+        "hex",
+        "txid",
+        "hash",
+        "size",
+        "version",
+        "locktime",
+        "vin",
+        "vout",
+        "blockhash",
+        "confirmations",
+        "time",
+        "blocktime"
+      ])
+      assert.isArray(result[0].vin)
+      assert.isArray(result[0].vout)
+    })
+
+    it(`should throw error on array size rate limit`, async () => {
+      try {
+        const dataMock =
+          "23213453b4642a73b4fc30d3112d72549ca153a8707255b14373b59e43558de1"
+        const data = []
+        for (let i = 0; i < 25; i++) data.push(dataMock)
+
+        const result = await bitbox.RawTransactions.getRawTransaction(data)
+
+        console.log(`result: ${util.inspect(result)}`)
+        assert.equal(true, false, "Unexpected result!")
+      } catch (err) {
+        assert.hasAnyKeys(err, ["error"])
+        assert.include(err.error, "Array too large")
+      }
     })
   })
 
-  describe("#sendRawTransaction", () => {
-    let sandbox: any
-    beforeEach(() => (sandbox = sinon.sandbox.create()))
-    afterEach(() => sandbox.restore())
+  describe("#decodeScript", () => {
+    it("should decode script for a single hex", async () => {
+      const hex =
+        "4830450221009a51e00ec3524a7389592bc27bea4af5104a59510f5f0cfafa64bbd5c164ca2e02206c2a8bbb47eabdeed52f17d7df668d521600286406930426e3a9415fe10ed592012102e6e1423f7abde8b70bca3e78a7d030e5efabd3eb35c19302542b5fe7879c1a16"
 
-    it("should send single raw transaction", async () => {
-      const data = "Error: transaction already in block chain"
+      const result = await bitbox.RawTransactions.decodeScript(hex)
+      //console.log(`result ${JSON.stringify(result, null, 2)}`)
 
-      const resolved = new Promise(r => r({ data: data }))
-      sandbox.stub(axios, "get").returns(resolved)
-
-      const result = await bitbox.RawTransactions.sendRawTransaction(
-        "020000000160d663961c63c7f0a07f22ec07b8f55b3935bfdbed8b1d8454916e8932fbf109010000006b4830450221008479fab4cfdcb111833d250a43f98ac26d43272b7a29cb1b9a0491eae5c44b3502203448b17253632395c29a7d62058bbfe93efb20fc8636ba6837002d464195aec04121029123258f7cdcd45b864066bcaa9b71f24d5ed1fa1dd36eaf107d8432b5014658ffffffff016d180000000000001976a91479d3297d1823149f4ec61df31d19f2fad5390c0288ac00000000"
-      )
-
-      assert.equal(data, result)
+      assert.hasAllKeys(result, ["asm", "type", "p2sh"])
     })
 
-    it("should send an array of raw transactions", async () => {
-      const data = "Error: transaction already in block chain"
+    // CT 2/20/19 - Waiting for this PR to be merged complete the test:
+    // https://github.com/Bitcoin-com/rest.bitcoin.com/pull/312
+    /*
+    it("should decode an array of tx hexes", async () => {
+      const hexes = [
+        "4830450221009a51e00ec3524a7389592bc27bea4af5104a59510f5f0cfafa64bbd5c164ca2e02206c2a8bbb47eabdeed52f17d7df668d521600286406930426e3a9415fe10ed592012102e6e1423f7abde8b70bca3e78a7d030e5efabd3eb35c19302542b5fe7879c1a16",
+        "4830450221009a51e00ec3524a7389592bc27bea4af5104a59510f5f0cfafa64bbd5c164ca2e02206c2a8bbb47eabdeed52f17d7df668d521600286406930426e3a9415fe10ed592012102e6e1423f7abde8b70bca3e78a7d030e5efabd3eb35c19302542b5fe7879c1a16"
+      ]
 
-      // Mock the http call to rest.bitcoin.com
-      nock(`${bitbox.RawTransactions.restURL}`)
-        .post((uri: string) => uri.includes(`/`))
-        .reply(200, { data: data })
+      const result = await bitbox.RawTransactions.decodeScript(hexes)
+      console.log(`result ${JSON.stringify(result, null, 2)}`)
+    })
+*/
+    /*
+    it(`should throw an error for improper single input`, async () => {
+      try {
+        const addr = 12345
 
-      const result = await bitbox.RawTransactions.sendRawTransaction([
-        "020000000160d663961c63c7f0a07f22ec07b8f55b3935bfdbed8b1d8454916e8932fbf109010000006b4830450221008479fab4cfdcb111833d250a43f98ac26d43272b7a29cb1b9a0491eae5c44b3502203448b17253632395c29a7d62058bbfe93efb20fc8636ba6837002d464195aec04121029123258f7cdcd45b864066bcaa9b71f24d5ed1fa1dd36eaf107d8432b5014658ffffffff016d180000000000001976a91479d3297d1823149f4ec61df31d19f2fad5390c0288ac00000000"
-      ])
+        await bitbox.RawTransactions.decodeRawTransaction(addr)
+        assert.equal(true, false, "Unexpected result!")
+      } catch (err) {
+        //console.log(`err: `, err)
+        assert.include(
+          err.message,
+          `Input must be a string or array of strings.`
+        )
+      }
+    })
+*/
+  })
 
-      assert.equal(data, result.data)
+  /*
+    Testing sentRawTransaction isn't really possible with an integration test,
+    as the endpoint really needs an e2e test to be properly tested. The tests
+    below expect error messages returned from the server, but at least test
+    that the server is responding on those endpoints, and responds consistently.
+  */
+  describe("sendRawTransaction", () => {
+    it("should send a single transaction hex", async () => {
+      try {
+        const hex =
+          "01000000013ba3edfd7a7b12b27ac72c3e67768f617fc81bc3888a51323a9fb8aa4b1e5e4a000000006a4730440220540986d1c58d6e76f8f05501c520c38ce55393d0ed7ed3c3a82c69af04221232022058ea43ed6c05fec0eccce749a63332ed4525460105346f11108b9c26df93cd72012103083dfc5a0254613941ddc91af39ff90cd711cdcde03a87b144b883b524660c39ffffffff01807c814a000000001976a914d7e7c4e0b70eaa67ceff9d2823d1bbb9f6df9a5188ac00000000"
+
+        await bitbox.RawTransactions.sendRawTransaction(hex)
+        //console.log(`result ${JSON.stringify(result, null, 2)}`)
+
+        assert.equal(true, false, "Unexpected result!")
+      } catch (err) {
+        //console.log(`err: ${util.inspect(err)}`)
+
+        assert.hasAllKeys(err, ["error"])
+        assert.include(err.error, "Missing inputs")
+      }
+    })
+
+    it("should send an array of tx hexes", async () => {
+      try {
+        const hexes = [
+          "01000000013ba3edfd7a7b12b27ac72c3e67768f617fc81bc3888a51323a9fb8aa4b1e5e4a000000006a4730440220540986d1c58d6e76f8f05501c520c38ce55393d0ed7ed3c3a82c69af04221232022058ea43ed6c05fec0eccce749a63332ed4525460105346f11108b9c26df93cd72012103083dfc5a0254613941ddc91af39ff90cd711cdcde03a87b144b883b524660c39ffffffff01807c814a000000001976a914d7e7c4e0b70eaa67ceff9d2823d1bbb9f6df9a5188ac00000000",
+          "01000000013ba3edfd7a7b12b27ac72c3e67768f617fc81bc3888a51323a9fb8aa4b1e5e4a000000006a4730440220540986d1c58d6e76f8f05501c520c38ce55393d0ed7ed3c3a82c69af04221232022058ea43ed6c05fec0eccce749a63332ed4525460105346f11108b9c26df93cd72012103083dfc5a0254613941ddc91af39ff90cd711cdcde03a87b144b883b524660c39ffffffff01807c814a000000001976a914d7e7c4e0b70eaa67ceff9d2823d1bbb9f6df9a5188ac00000000"
+        ]
+
+        const result = await bitbox.RawTransactions.sendRawTransaction(hexes)
+        console.log(`result ${JSON.stringify(result, null, 2)}`)
+      } catch (err) {
+        // console.log(`err: ${util.inspect(err)}`)
+
+        assert.hasAllKeys(err, ["error"])
+        assert.include(err.error, "Missing inputs")
+      }
+    })
+
+    it(`should throw an error for improper single input`, async () => {
+      try {
+        const addr = 12345
+
+        await bitbox.RawTransactions.sendRawTransaction(addr)
+        assert.equal(true, false, "Unexpected result!")
+      } catch (err) {
+        //console.log(`err: `, err)
+        assert.include(
+          err.message,
+          `Input hex must be a string or array of strings`
+        )
+      }
+    })
+
+    it(`should throw error on array size rate limit`, async () => {
+      try {
+        const dataMock =
+          "01000000013ba3edfd7a7b12b27ac72c3e67768f617fc81bc3888a51323a9fb8aa4b1e5e4a000000006a4730440220540986d1c58d6e76f8f05501c520c38ce55393d0ed7ed3c3a82c69af04221232022058ea43ed6c05fec0eccce749a63332ed4525460105346f11108b9c26df93cd72012103083dfc5a0254613941ddc91af39ff90cd711cdcde03a87b144b883b524660c39ffffffff01807c814a000000001976a914d7e7c4e0b70eaa67ceff9d2823d1bbb9f6df9a5188ac00000000"
+        const data = []
+        for (let i = 0; i < 25; i++) data.push(dataMock)
+
+        const result = await bitbox.RawTransactions.sendRawTransaction(data)
+
+        console.log(`result: ${util.inspect(result)}`)
+        assert.equal(true, false, "Unexpected result!")
+      } catch (err) {
+        assert.hasAnyKeys(err, ["error"])
+        assert.include(err.error, "Array too large")
+      }
     })
   })
 })
